+++ recepcion_pedido/recepcion.html
@@ -387,5 +387,7 @@
         </div>
       </main>
     </div>
+
+    <script src="recepcion.js" type="module"></script>
   </body>
 </html>
